Add files via upload
This application is user friendly and it will restrict the user
from accessing other user’s data. This application will be
consume small amount of power compare to other application
and It will help user in viewing his data and privileges. And it
will also help the administrator to handle all the changes.